Finalizacion flujo historia 12
faltan las historias de usuario

+++ src/main/java/com/emazon/ApiCart/Domain/Spi/StockFeignPort.java
@@ -2,5 +2,5 @@ package com.emazon.ApiCart.Domain.Spi;
 
 
 public interface StockFeignPort {
-    Long checkStock(Long itemId,long quantity);
+    double checkStock(Long itemId,long quantity);
 }
